Refactor HoodieCommonUtils to make code more reasonable
